switch back to built-in mocha+chai test page
Now that subject is being served from the standard build file (dist/),
and mocha-given is being served from node_modules, we don't need custom
script tags in the test page. And the browser-helper sets mocha-given as
the mocha UI so we don't need that in the test page either. That gets us
back to the OOTB mocha+chai test page that ships with testem.