fix version script lifecycle
`preversion` script has access to the old version (not the bumped
version), so committing a build here with the $npm_package_version
variable would be comitted with the wrong version number

the package.json version is bumped immediately prior to the `version`
lifecycle script. this is typically where building and git-adding should
occur. This script has access to the new (bumped) version, so any
version references in the generated files will be correct

(Can't do file generation in `postversion` because that script is run
*after* the commit/tag)

Lastly, npm does the git commit+tag for us, we don't need to execute a
git commit. It's done immediately after the `version` script (just before
`postversion`), and will commit (along with package.json) any files
currently in the index; which for our purposes is dist/* (added to index
in the `version` script