add envify transform
the envify transform will replace any `process.env` usage with the
static values at build time. If the env variable isn't set, nothing
will be substituted so it will still pick up the runtime value.

This means that any build-time env vars can be statically added to the
build without browserify needing to include the browserify-core to get
`process` at runtime.

Specifically for us, we're grabbing the version out of
npm_package_version so that we can skip the package.json require