Add unreachable-branch-transform
This transform eliminates dead-code branches.

Specifically, we use this so that once the npm_package_version variable
has been substituted with a static value (via envify), the || conditional is now
satisfied so the package.json require becomes dead code. This transform
removes the require so browserify doesn't need to include package.json
in the bundle.