explicit compile before publish
coffee already implicitly compiled before published because we use
npm-version (which does a build which does a compile) is what kicks off
a publish step. However, publish is also invoked whenever a package is
installed at the top level. (Such as when `npm install` is run inside of
a cloned repo) Compilation is necessary in this case so that a single
npm-install will result in a requireable package. This is precisely why
npm runs publish on `npm install` (with no args)