Make it pass. Add a new universal config option.
Just keep adding ", config" until the pain goes away.

So this should grease the skid to other options when
determining a match by just passing it all the way through from
the user down into the arg matcher function. It's a bit of a blunt
instrument, but I can't think of a simpler way to do it.