Fix the over-wide stubbing count
Previously, the state of the test double prior
to a thenReturn(1,2,3) would mean that if the td
had been invoked previously, that callCount would
count against the later sequenced stubbing.

This shouldn't have come up in practice often or ever, 
but if it ever did it'd almost certainly be considered wrong by
the tester.