split 'create' into 'function' and 'object'
If we add proxy objects it's going to be too damn awkward otherwise

What is a "test double"? is it a function or an object? the correct 
answer should be it could be either, so let's not pretend one is the 
'real' test double and the other is just an object on it or a function
of it