Add webpack example
This example has a basic webpack configuration that attempts to build a
test bundle.  There’s also an `npm run test:debug` script to allow
running the tests in the browser (at `localhost:8081`).

The tests can’t be run from the command line, but this example
demonstrates the issue reported in #77.  Once that issue is resolved,
the example should build correctly, and the tests should run in the
browser.