Make the module compatible with Weback
So, it turns out:

1. Webpack does support the `browser` field, we just
had to replace quibble with an empty shim, which we did
not have to do for browserify

2. Webpack doesn't have a way to load json OOTB; it's a bummer
it doesn't use the node module resolution algorithm as a baseline,
especially since their json-loader is literally 4 lines long, but

3. I created a script that literally echoes out a version module
at `npm version`-time. Not ideal but a one-time cost to us.