Improves warning when verifying a stub
After a verification is made, a warning can be produced to signal that
the verification call matches a stub on the same testdouble. This
behavior is improved to test the verification against the stubbing using
the stubs matchers and config options.

This closes: stub also verified warning when using ignoreExtraArgs #181
completed with @tinney