undo current implementation
My issue with this implementation is that it puts the responsibility
of making the double in the `td.replace` feature, but 
that responsibility belongs to `td.function`. So I think the
right place to implement it is as a broadened support for
`td.function` to accept not only an (optional) function name,
but also an existing function for which it shall be imitated
(e.g. name applied & function properties doubled)