Get a start on killing default type extension.
The biggest ramification of this is that static & instance attributes
defined for the type will not be copied over. Probably nothing to be
done about most instance attributes (since the whole point of this is to not
call the real constructor, and that's when all non-contrived instance
attrs are set).

Plan is to shallow copy over non-function properties from the type and
the prototype and see where that gets us.