Test testdouble.js with testdouble.js!
I've never been happy with this library's "unit" test suite because it's
too extrinsic, poking at things only through the top-level exposed td.*
API methods. This is a fine strategy for smaller libraries, but on
larger libraries like this one, it encourages large units and extract
refactors in the src code tend to be more arbitrary and less symmetrical
to the test files, which increases onboarding time for others and adds
some indirection when trying to find the spot in the code that a test is
specifying.

This was done for two reasons at the time, both of which I now regret:

1. testdouble.js didn't exist yet, so doing my brand of Discovery
Testing would be hard.
2. I didn't trust Browserify to work, so I went out of my way at first
to test the built Browserified lib extrinsically in a browser
environment to make sure the build was right, which meant I couldn't
`require` any internals inside the test suite.

As a result, the test suite we have is not a unit suite so much as an
exhaustive functional test suite. This may have made sense at the time,
but Node.js & Browsers have reached a level of maturity and parity that
non-DOM APIs like td.js should be able to run unit tests in Node and
have relative confidence that the same stuff would also pass in browsers
(meaning it no longer makes sense to bend over backwards to run the unit
tests in a native browser environment).

So what this PR is getting started is a *BRAND NEW TEST SUITE* that will
try to dogfood testdouble and "rediscovery test" it from the start,
which will require pretty drastic refactoring and reimagining of the
code structure to feature smaller units, better names, and a strict
hierarchical internal dependency graph.

It accomplishes this by using the mocha/coffeescript/mocha-given functional test
suite as a refactoring backstop (to be deleted after we're done) while
we write a brand new unit test suite in Test Double®'s own teenytest
test framework & a prior known version of testdouble.js (3.0.0 at this
time). That means that testdouble.js may become the first self-hosted
mocking library I know of—neato!