Use this value object
Now that we have a value for Double metadata, return the fake dangling
on it to the user.

Note that this tweak likely signs us up for mutating-the-hell out of these
double objects. I thought hard about this, but because they're
long-lived and effectively scratch-pads, there's not much way around it
without inserting tons of wrap-and-replace for, say, every time a call
is registered. JS isn't a pure functional language, so I guess the best
we can do here without taking immutability to an awkward extreme is to
sequester that mutability into these values and out of the line of fire
of the library logic.