Use a parent/ancestor data to print names
This was a doozy. In order to avoid mucking with the td.func() API like
the current impl does (see __optionalName field), I wanted to be able to
change names after the fact. Moreover, I wanted child doubles to retain
names from their parents, which had previously been hiding in a
closure-scoped variable and which we lose by breaking things up.

This forced me to slam more metadata into the Double object. I'm not
thrilled that I'm mutating the Double after it's created (e.g.
addChild()), but that was necessary to reuse the current create() method
without burdening it with knowledge of parents. That said, the metadata
itself could be useful for implementing a much better td.explain()
function someday, one that knows all about the double's heritage