Remove extendWhenReplacingConstructors option
The improvements in the new imitate function sidestep the babel class
extension issue that prompted the introduction of this option. All fakes
will pass instanceof checks as a matter of course now.

Of course, this means that a major bump will be necessary, since that
config option will have been removed and anyone setting it will blow up…