Pair with @andrewvida to refactor imitate()
This refactor started by teasing the tangled function into its separate
roles, then pushing the behavior down into 3 new units.

The unit test for the refactored thing is particularly interesting,
because it required use of argument captors to trigger the downstream
behavior. This was necessary because the method was recursive, and we
needed to invoke the subject twice, once indirectly. It's hard to find
good examples of argument captors, so here you go.