Fix bug caused by things with undefined prototypes
A user was doing td.replace('aws-sdk'), which is massive and contains a
lot of really screwy prototype manipulation that apparently has at least
one type with an undefined in the prototype chain of one of its
functions.

This guards against that. The test isn't perfect because I wasn't even
able to reproduce the same prototype chain nonsense