Rewrite the td.function() rewrite
Key insight here was that the public API really shouldn't be the same
as the private API, because we want to encourage passing around Double
objects internally and only expose the fake externally. Additionally,
the imitation module is doing a lot of the heavy lifting now and this
isn't nearly as complicated as it was once going to be.