With double rewrite, imitate will use private API.
I suspect we'll move to a mode where every single layer of an imitation
is mapped to a Double value, regardless of whether `fake` is a clone or
a test double function. it'll allow us to say quite a lot more when 
td.explain()'ing. 

However, that's not something we have to solve right away (pre-relaunch)