Remove satisfy() from stubbing store
I could not for the life of me figure out why app code was being invoked
from a value object. Obviously it's a thing I (tell people to) avoid
doing, but as I dove in I realized that this stupid unit test was also
depending on this dummy implementation of satisfy() working in order
to also pass.

Not only was that unhelpful incidental coverage, it was obnoxious b/c
I needed to delete that implementation to start driving out an actual
satisfy() solution, and I didn't want to stare at this unrelated error.