Step 8: Make the third test pass
I expected this test to pass without changes, but it didn't!
To my surprise, destructured parameter assignment explodes
if the right side is null or undefined (this gives a good
error message in native v8, but when babelified the message
is not helpful).

Now, a decision: leave this gross `|| {}`, adding to the branching
complexity of the unit to mask over this language quibble, OR refactor
things around to push that complexity onto ensureDemonstration, so it
can effectively serve as the wrapper, throwing instead of returning a 
null or a default.