Change proposed solution
The proposed solution didn't work, because the `arguments.length` was
always 2 since it was an internal private method. Since this is about to
be rewritten anyway I just tacked on a stupid third arg that measures 
arg length on the public function.

I also removed the extra error message since the original message is 
actually probably clearer. I highly doubt anyone is intentionally 
relying on this, so a compat notice isn't very necessary IMO