Add awareness of testdouble-jest to td.replace
If a user calls td.replace for a module from a jest test, it will
delegate to a new jest-specific module faker. If testdouble-jest is
not initialized (as detected by an absence of `td.mock`), then the
lib will complain, to hopefully make folks aware of testdouble-jest
in the first place. 

When td.mock is detected, it will just translate the `td.replace` 
contract, which actually returns the fake. This should be more portable.

There are a few warts, namely that I can't figure out how to resolve a
path, so I have to attempt to require manually stubbed items in a 
try-catch (sort of defeating the purpose of a stub override) to know
whether to flip Jest's virtual bit or not. (If the dep is real and 
virtual is set to true, the real dep apparently wins)