Eliminate unit test browser build
Up to now, we've always run our regression suite first under Node and 
then under Phantom (and now Chrome). In 3 years this has cost me about
20 hours in build configuration headache and has caught maybe one bug.

As we look ahead to simplifying the overall build, I'm eliminating this
again. Once we're done cleaning house, I'd be open to reintroducing this
and suspect it'd take a different form