A way too clever way of getting a global run
Wrote a teenytest plugin to basically filter each test into "unit" vs
not. 

If the test lives in test/unit/…, then `td` is set to the npm
release so they can use a known good version of the lib for mocking.

If the test lives in test/safe then `td` is set to `src/index` so
the working copy can be tested.

This lets us run a single combined `npm test` and also simplifies our
code coverage report into one combined one