Add support for containing error messages.
Since two Errors are never rightly equal, adds contains() support if
their messages are the same. In fact, so often a message has superfluous
stuff at the beginning and end, "contains" should pass even if what is
specified by the test is a substring.

This should make it a little easier for testing invocations that receive
an error argument