ci - teenytest-proxy Windows support
npm installs packages differently based on the platform.  The `.bin`
entry the proxy was using is a symlink on UNIX type platforms but may have
three different forms on WIndows.  The actual installed package will
now  be required directly.  This is still a temporary patch pending
fixes for the issues identified in the proxy file.

Signed-off-by: Chris. Webster <chris@webstech.net>